Merge recent updates  into consenus branch

+++ neo/Prompt/Commands/Bootstrap.py
@@ -8,42 +8,40 @@ import shutil
 import os
 
 
-def BootstrapBlockchain():
+def BootstrapBlockchainFile(target_dir, download_file, require_confirm=True):
 
-    current_chain_dir = settings.LEVELDB_PATH
-
-    bootstrap_file = settings.BOOTSTRAP_FILE
-
-    if bootstrap_file is None:
+    if download_file is None:
         print("no bootstrap file specified.  Please update your configuration file.")
         sys.exit(0)
 
-    print("This will overwrite any data currently in %s.\nType 'confirm' to continue" % current_chain_dir)
+    if require_confirm:
+        print("This will overwrite any data currently in %s.\nType 'confirm' to continue" % target_dir)
+        confirm = prompt("[confirm]> ", is_password=False)
+        if confirm == 'confirm':
+            return do_bootstrap(download_file, target_dir)
+    else:
 
-    confirm = prompt("[confirm]> ", is_password=False)
-
-    if confirm == 'confirm':
-        return do_bootstrap()
+        return do_bootstrap(download_file,
+                            target_dir,
+                            tmp_file_name=os.path.join(settings.DATA_DIR_PATH, 'btest.tar.gz'),
+                            tmp_chain_name='btestchain')
 
     print("bootstrap cancelled")
     sys.exit(0)
 
 
-def do_bootstrap():
+def do_bootstrap(bootstrap_file, destination_dir, tmp_file_name=None, tmp_chain_name='tmpchain'):
 
-    bootstrap_file = settings.BOOTSTRAP_FILE
-    destination_dir = settings.LEVELDB_PATH
+    if tmp_file_name is None:
+        tmp_file_name = os.path.join(settings.DATA_DIR_PATH, 'bootstrap.tar.gz')
 
     success = False
 
     print('will download file %s ' % bootstrap_file)
     print('')
-    tmp_file_name = './Chains/bootstrap.tar.gz'
-    tmp_chain_name = 'tmpchain'
 
     try:
         response = requests.get(bootstrap_file, stream=True)
-
         response.raise_for_status()
 
         # Total size in bytes.
@@ -60,11 +58,9 @@ def do_bootstrap():
         print("download complete")
 
         if os.path.exists(destination_dir):
-
             try:
                 shutil.rmtree(destination_dir)
             except Exception as e:
-
                 print("coludnt remove existing dir: %s %s" % (e, destination_dir))
                 sys.exit(0)
 
@@ -95,11 +91,9 @@ def do_bootstrap():
         print("Could not download: %s " % e)
 
     finally:
-
-        print("cleaning up %s " % tmp_file_name)
-#        os.remove(tmp_file_name)
         print("cleaning up %s " % tmp_chain_name)
-        shutil.rmtree(tmp_chain_name)
+        if os.path.exists(tmp_chain_name):
+            shutil.rmtree(tmp_chain_name)
 
     if success:
         print("Successfully downloaded bootstrap chain!")
